Conflicts:
	src/main/groovy/com/blackducksoftware/integration/hub/detect/DetectProperties.groovy

+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/DetectProjectManager.groovy
@@ -22,6 +22,7 @@
  */
 package com.blackducksoftware.integration.hub.detect
 
+import org.apache.commons.io.FilenameUtils
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -35,13 +36,13 @@ import com.blackducksoftware.integration.hub.bdio.simple.model.BdioBillOfMateria
 import com.blackducksoftware.integration.hub.bdio.simple.model.BdioComponent
 import com.blackducksoftware.integration.hub.bdio.simple.model.BdioExternalIdentifier
 import com.blackducksoftware.integration.hub.bdio.simple.model.BdioProject
+import com.blackducksoftware.integration.hub.bdio.simple.model.DependencyNode
 import com.blackducksoftware.integration.hub.bdio.simple.model.SimpleBdioDocument
 import com.blackducksoftware.integration.hub.detect.bomtool.BomTool
 import com.blackducksoftware.integration.hub.detect.bomtool.output.DetectCodeLocation
 import com.blackducksoftware.integration.hub.detect.bomtool.output.DetectProject
 import com.blackducksoftware.integration.hub.detect.hub.HubSignatureScanner
 import com.blackducksoftware.integration.hub.detect.type.BomToolType
-import com.blackducksoftware.integration.hub.detect.util.ProjectInfoGatherer
 import com.blackducksoftware.integration.util.ExcludedIncludedFilter
 import com.blackducksoftware.integration.util.IntegrationEscapeUtil
 import com.google.gson.Gson
@@ -52,9 +53,6 @@ class DetectProjectManager {
 
     @Autowired
     DetectConfiguration detectConfiguration
-
-    @Autowired
-    ProjectInfoGatherer projectInfoGatherer
 
     @Autowired
     BdioPropertyHelper bdioPropertyHelper
@@ -73,6 +71,9 @@ class DetectProjectManager {
 
     @Autowired
     HubSignatureScanner hubSignatureScanner
+
+    @Autowired
+    IntegrationEscapeUtil integrationEscapeUtil
 
     private boolean foundAnyBomTools
 
@@ -115,12 +116,12 @@ class DetectProjectManager {
         }
 
         //if none of the bom tools could determine a project/version, use some reasonable defaults
-        detectProject.projectName = projectInfoGatherer.getProjectName(detectConfiguration.sourcePath, detectProject.projectName)
-        detectProject.projectVersionName = projectInfoGatherer.getProjectVersionName(detectProject.projectVersionName, detectProject.projectVersionHash)
+        detectProject.projectName = getProjectName(detectProject.projectName)
+        detectProject.projectVersionName = getProjectVersionName(detectProject.projectVersionName, detectProject.projectVersionHash)
 
         if (!foundAnyBomTools) {
             logger.info("Could not find any tools to run - will register ${detectConfiguration.sourcePath} for signature scanning of ${detectProject.projectName}/${detectProject.projectVersionName}")
-            hubSignatureScanner.registerDirectoryToScan(detectConfiguration.sourceDirectory)
+            hubSignatureScanner.registerPathToScan(detectConfiguration.sourceDirectory)
         }
 
         detectProject
@@ -128,41 +129,72 @@ class DetectProjectManager {
 
     public List<File> createBdioFiles(DetectProject detectProject) {
         List<File> bdioFiles = []
+        final String safeProjectName = integrationEscapeUtil.escapeForUri(detectProject.projectName)
+        final String safeVersionName = integrationEscapeUtil.escapeForUri(detectProject.projectVersionName)
+
+        File aggregateBdioFile = null
+        final SimpleBdioDocument aggregateBdioDocument = null
+        if (detectConfiguration.aggregateBomName) {
+            aggregateBdioDocument = createAggregateSimpleBdioDocument(detectProject)
+            final String filename = "${integrationEscapeUtil.escapeForUri(detectConfiguration.aggregateBomName)}.jsonld"
+            aggregateBdioFile = new File(detectConfiguration.getOutputDirectory(), filename)
+            if (aggregateBdioFile.exists()) {
+                aggregateBdioFile.delete()
+            }
+        }
+
         detectProject.detectCodeLocations.each {
-            File createdBdioFile = createBdioFile(detectProject, it)
-            bdioFiles.add(createdBdioFile)
+            if (detectConfiguration.aggregateBomName) {
+                aggregateBdioDocument.components.addAll(dependencyNodeTransformer.addComponentsGraph(aggregateBdioDocument.project, it.dependencies))
+            } else {
+                final SimpleBdioDocument simpleBdioDocument = createSimpleBdioDocument(detectProject, it)
+                final String filename = "${it.bomToolType.toString()}_${safeProjectName}_${safeVersionName}_bdio.jsonld"
+                final File outputFile = new File(detectConfiguration.getOutputDirectory(), filename)
+                if (outputFile.exists()) {
+                    outputFile.delete()
+                }
+                final File createdBdioFile = writeSimpleBdioDocument(outputFile, simpleBdioDocument)
+                bdioFiles.add(createdBdioFile)
+            }
+        }
+
+        if (aggregateBdioFile != null && aggregateBdioDocument != null) {
+            writeSimpleBdioDocument(aggregateBdioFile, aggregateBdioDocument)
         }
 
         bdioFiles
     }
 
-    private File createBdioFile(DetectProject detectProject, DetectCodeLocation detectCodeLocation) {
-        String projectName = detectProject.projectName
-        String projectVersionName = detectProject.projectVersionName
-        String codeLocationName = projectInfoGatherer.getCodeLocationName(detectCodeLocation.bomToolType, detectCodeLocation.sourcePath, projectName, projectVersionName)
+    private SimpleBdioDocument createAggregateSimpleBdioDocument(DetectProject detectProject) {
+        createSimpleBdioDocument(detectProject, '', detectProject.projectName, bdioPropertyHelper.createExternalIdentifier('', detectProject.projectName), [] as Set)
+    }
 
-        final IntegrationEscapeUtil escapeUtil = new IntegrationEscapeUtil()
-        final String safeProjectName = escapeUtil.escapeForUri(projectName)
-        final String safeVersionName = escapeUtil.escapeForUri(projectVersionName)
-        final String filename = String.format("%s_%s_%s_bdio.jsonld", detectCodeLocation.bomToolType.toString(), safeProjectName, safeVersionName)
-        final File outputFile = new File(detectConfiguration.getOutputDirectory(), filename)
-        if (outputFile.exists()) {
-            outputFile.delete()
-        }
-
-        BdioBillOfMaterials bdioBillOfMaterials = bdioNodeFactory.createBillOfMaterials(codeLocationName, projectName, projectVersionName)
-
+    private SimpleBdioDocument createSimpleBdioDocument(DetectProject detectProject, DetectCodeLocation detectCodeLocation) {
+        final String codeLocationName = getCodeLocationName(detectCodeLocation.bomToolType, detectCodeLocation.sourcePath, detectProject.projectName, detectProject.projectVersionName)
         final String projectId = detectCodeLocation.bomToolProjectExternalId.createDataId()
         final BdioExternalIdentifier projectExternalIdentifier = bdioPropertyHelper.createExternalIdentifier(detectCodeLocation.bomToolProjectExternalId)
+
+        createSimpleBdioDocument(detectProject, codeLocationName, projectId, projectExternalIdentifier, detectCodeLocation.dependencies)
+    }
+
+    private SimpleBdioDocument createSimpleBdioDocument(DetectProject detectProject, String codeLocationName, String projectId, BdioExternalIdentifier projectExternalIdentifier, Set<DependencyNode> dependencies) {
+        final String projectName = detectProject.projectName
+        final String projectVersionName = detectProject.projectVersionName
+
+        final BdioBillOfMaterials bdioBillOfMaterials = bdioNodeFactory.createBillOfMaterials(codeLocationName, projectName, projectVersionName)
         final BdioProject project = bdioNodeFactory.createProject(projectName, projectVersionName, projectId, projectExternalIdentifier)
 
-        final List<BdioComponent> bdioComponents = dependencyNodeTransformer.addComponentsGraph(project, detectCodeLocation.dependencies)
+        final List<BdioComponent> bdioComponents = dependencyNodeTransformer.addComponentsGraph(project, dependencies)
 
-        SimpleBdioDocument simpleBdioDocument = new SimpleBdioDocument()
+        final SimpleBdioDocument simpleBdioDocument = new SimpleBdioDocument()
         simpleBdioDocument.billOfMaterials = bdioBillOfMaterials
         simpleBdioDocument.project = project
         simpleBdioDocument.components = bdioComponents
 
+        simpleBdioDocument
+    }
+
+    private File writeSimpleBdioDocument(File outputFile, SimpleBdioDocument simpleBdioDocument) {
         final BdioWriter bdioWriter = new BdioWriter(gson, new FileOutputStream(outputFile))
         try {
             bdioWriter.writeSimpleBdioDocument(simpleBdioDocument)
@@ -172,5 +204,47 @@ class DetectProjectManager {
         logger.info("BDIO Generated: " + outputFile.getAbsolutePath())
 
         outputFile
+    }
+
+    String getProjectName(final String defaultProjectName) {
+        String projectName = defaultProjectName?.trim()
+
+        if (detectConfiguration.getProjectName()) {
+            projectName = detectConfiguration.getProjectName()
+        } else if (!projectName && detectConfiguration.sourcePath) {
+            String finalSourcePathPiece = extractFinalPieceFromSourcePath(detectConfiguration.sourcePath)
+            projectName = finalSourcePathPiece
+        }
+
+        projectName
+    }
+
+    String getProjectVersionName(final String defaultVersionName, final String bomToolFileHash) {
+        String projectVersion = defaultVersionName?.trim()
+
+        if (detectConfiguration.getProjectVersionName()) {
+            projectVersion = detectConfiguration.getProjectVersionName()
+        } else if (!projectVersion) {
+            projectVersion = 'Detect Unknown Version'
+        }
+
+        projectVersion
+    }
+
+    String getCodeLocationName(final BomToolType bomToolType, final String sourcePath, final String projectName, final String projectVersion) {
+        String codeLocation = detectConfiguration.getProjectCodeLocationName()
+        if (!codeLocation?.trim()) {
+            String finalSourcePathPiece = extractFinalPieceFromSourcePath(sourcePath)
+            codeLocation = String.format('%s/%s/%s', finalSourcePathPiece, projectName, projectVersion)
+        }
+        return String.format('%s/%s Hub Detect Export', bomToolType.toString(), codeLocation)
+    }
+
+    private String extractFinalPieceFromSourcePath(String sourcePath) {
+        if (sourcePath == null || sourcePath.length() == 0) {
+            return ''
+        }
+        String normalizedSourcePath = FilenameUtils.normalizeNoEndSeparator(sourcePath, true)
+        normalizedSourcePath[normalizedSourcePath.lastIndexOf('/') + 1..-1]
     }
 }
